Spring 프레임워크 적용 및 서버 코드 구조 변경
이 커밋에서는 Spring 프레임워크를 적용하고, 서버 코드의 구조를 변경하였습니다. 이를 통해 코드의 유지보수성과 확장성을 향상시켰습니다.

주요 변경사항:
1. `build.gradle.kts` 파일에서 Spring 프레임워크 의존성을 추가하였습니다.
2. `App.java`에서 서버 실행 코드를 `Server.java`로 이동하였습니다. 이를 통해 `App.java`는 Spring 컨텍스트를 초기화하는 역할만 수행하게 되었습니다.
3. `Server.java`를 새로 생성하여 서버 실행 코드를 관리하게 하였습니다. 이 클래스는 Spring 빈으로 등록되어 `App.java`에서 관리됩니다.
